% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMM.R
\name{PMMLogLik}
\alias{PMMLogLik}
\title{Calculate the PMM log-likelihood for a given tree, data and model parameters}
\usage{
PMMLogLik(x, tree, x0, sigma2, sigmae2, ord = reorder(tree, order =
  "postorder", index.only = TRUE))
}
\arguments{
\item{x}{a numerical vector of size N, where N is the number of tips in tree}

\item{tree}{a phylo object}

\item{x0, sigma2, sigmae2}{parameters of the PMM:
\describe{
\item{x0}{value at the root of tree excluding whte noise;}
\item{sigma2}{unit-time variance increment of the heritable component;}
\item{sigmae2}{variance of the non-heritable component.}
}}

\item{ord}{an integer vector denoting the pruning order. This should be the 
result from calling `reorder(tree, order = "postorder", index.only = TRUE)`, 
which is also set as a default value. Can be passed as argument to speed-up 
the calculation.}
}
\value{
the log-likelihood value.
}
\description{
Calculate the PMM log-likelihood for a given tree, data and model parameters
}
