% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocForRcppModules.R
\name{PMMUsingSPLITT__AbcPMM__AlgorithmType}
\alias{PMMUsingSPLITT__AbcPMM__AlgorithmType}
\alias{Rcpp_PMMUsingSPLITT__AbcPMM__AlgorithmType-class}
\title{\code{TraversalAlgorithm}-type used in \code{AbcPMM}}
\description{
\code{TraversalAlgorithm}-type used in \code{AbcPMM}
}
