% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocForRcppModules.R
\name{PMMUsingSPLITT__AbcPMM__TraversalAlgorithm}
\alias{PMMUsingSPLITT__AbcPMM__TraversalAlgorithm}
\alias{Rcpp_PMMUsingSPLITT__AbcPMM__TraversalAlgorithm-class}
\title{Base class for \code{PMMUsingSPLITT::AbcPMM::AlgorithmType}}
\description{
Base class for \code{PMMUsingSPLITT::AbcPMM::AlgorithmType}
}
\section{Slots}{

\describe{
\item{\code{algorithm}}{algorithm.}
}}

