% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocForRcppModules.R
\name{PMMUsingSPLITT__TraversalTaskAbcPMM}
\alias{PMMUsingSPLITT__TraversalTaskAbcPMM}
\alias{Rcpp_PMMUsingSPLITT__TraversalTaskAbcPMM-class}
\title{Rcpp module for the \code{TraversalTaskAbcPMM}-class}
\description{
Rcpp module for the \code{TraversalTaskAbcPMM}-class
}
